+++ public/changelog.md
@@ -1,3 +1,6 @@
+# Update (05/30/2022): Basic tank skins!
+- To use one of this skins, include "[farm]". "[donut]" or "[sharp]" somewhere in your name!
+
 # Update (05/30/2022): Everything is not broken
 - I fixed it! Still no reskins however, im looking
 for a good way to implement those.